High Prio Interupt + ISR in Assembler, dadurch Prio 5 möglich, ergibt ein stabileres Bild + nutzt jetzt beide Cores: core0 für OSD und System, core1 für capture + die main.c gesplittet ++ capture.c : alles was mit dem Einlesen des Bildes zu tun hat ++ highint5.S : der ISR, gehört mit zum capture ++ osd.c : alles für das Setup-Menü ++ pins.h : die Pin-Definition. Als extra Datei, damit die in die highint5.S integriert werden kann ++ vga.c : alles für die VGA-Ausgabe
