+++ components/vga-adapter/capture.c
@@ -1,0 +1,211 @@
+
+// alles für die Abtastung der Video-Signale vom ABG mit SPI vorbereiten
+void setup_abg()
+{
+	// Buffer holen
+	uint8_t* ABG_PIXBUF1 = heap_caps_malloc(4096, MALLOC_CAP_DMA | MALLOC_CAP_32BIT | MALLOC_CAP_INTERNAL);
+	uint8_t* ABG_PIXBUF2 = heap_caps_malloc(4096, MALLOC_CAP_DMA | MALLOC_CAP_32BIT | MALLOC_CAP_INTERNAL);
+	ABG_DMALIST = heap_caps_malloc(32, MALLOC_CAP_DMA | MALLOC_CAP_32BIT | MALLOC_CAP_INTERNAL);
+	PIXEL_STEP_LIST = heap_caps_malloc(640, MALLOC_CAP_DMA | MALLOC_CAP_32BIT | MALLOC_CAP_INTERNAL);
+
+	// Pin-Konfiguration
+	spi_bus_config_t buscfg=
+	{
+		.data0_io_num = PIN_NUM_ABG_VIDEO1,  // Farb-Signal
+    	.data1_io_num = PIN_NUM_ABG_VIDEO2,  // Farb-Signal
+    	.data2_io_num = PIN_NUM_ABG_BSYNC1,  // wir zeichnen auch das BSYN auf, daran können wir später die Spalten ausrichten
+    	.max_transfer_sz = 4092,
+    	.flags = SPICOMMON_BUSFLAG_MASTER | SPICOMMON_BUSFLAG_GPIO_PINS,
+    };
+	ESP_ERROR_CHECK(spi_bus_initialize(SPI2_HOST, &buscfg, SPI_DMA_CH_AUTO));
+
+	// SPI-Konfiguration
+	spi_device_interface_config_t devcfg=
+    {
+    	.cs_ena_pretrans = 0,
+		.cs_ena_posttrans = 0,
+        .clock_speed_hz = SPI_MASTER_FREQ_40M, // Abtast-Frequenz 40MHz - das ist mehr als wir eigendlich brauchen,
+        .flags = SPI_DEVICE_HALFDUPLEX,        // aber mit weniger gäbe es Probleme mit der Pixelsynchronisation
+	    .queue_size=1,
+#ifdef DEBUG
+		.spics_io_num = PIN_NUM_DEBUG_SPI,     // zur Fehlersuche: an diesem Pin kann man sehen, wann das SPI aktiv ist
+#endif
+    };
+	spi_device_handle_t spi;
+    ESP_ERROR_CHECK(spi_bus_add_device(SPI2_HOST, &devcfg, &spi));
+
+    // Transfer-Konfiguration
+	spi_transaction_t transfer =
+	{
+	    .flags = SPI_TRANS_MODE_OCT,           // 8 Bit gleichzeitig einlesen. Es würde auch 4 reichen, aber dann wird die Auswertung langsamer und komplizierter
+	    .length = 0,                           // nix ausgeben...
+	    .rxlength = 1000,                      // Platzhalter, wird später aus dem hsync-timing errechnet
+	    .rx_buffer = ABG_PIXBUF1,
+	};
+
+	// Transfer starten. Wir brauchen die Daten hier noch nicht, aber der Transfer setzt alle Parameter in dem SPI-Kontroller
+	ESP_ERROR_CHECK(spi_device_queue_trans(spi, &transfer, 0));
+
+	// Aus dieser Tabelle nimmt der DMA-Kontroller die Speicheraddressen
+	ABG_DMALIST[0] = 0x80ffffff;
+	ABG_DMALIST[1] = (int)ABG_PIXBUF1;
+	ABG_DMALIST[2] = (int)&ABG_DMALIST[4];
+	ABG_DMALIST[4] = 0x80ffffff;
+	ABG_DMALIST[5] = (int)ABG_PIXBUF2;
+	ABG_DMALIST[6] = (int)&ABG_DMALIST[0];
+
+	// normalerweise nimmt der SPI-Kontroller bei diesem Projekt den DMA-Kanal 1. Sicher sein kann man sich da nicht - also überprüfen.
+	assert(REG_READ(GDMA_IN_PERI_SEL_CH1_REG) == 0);
+
+	// DMA-Kontroller 1 auf unseren Puffer umstellen und starten
+	REG_SET_BIT(GDMA_IN_LINK_CH1_REG, GDMA_INLINK_STOP_CH1);
+	REG_SET_BIT(GDMA_IN_CONF1_CH1_REG, GDMA_IN_CHECK_OWNER_CH1);
+	REG_SET_FIELD(GDMA_IN_LINK_CH1_REG, GDMA_INLINK_ADDR_CH1, (int)ABG_DMALIST);
+	REG_SET_BIT(GDMA_IN_LINK_CH1_REG, GDMA_INLINK_START_CH1);
+
+	// jetzt noch die restlichen GPIO's konfigurieren
+	// SPI und Interrupt gleichzeitig auf dem selben PIN klappt nicht, wir nehmen deshalb 2 getrennte Pins außen verbunden.
+	gpio_config_t pincfg =
+	{
+	    .pin_bit_mask = 1ULL << PIN_NUM_ABG_BSYNC2,
+	    .mode = GPIO_MODE_INPUT,
+	    .pull_up_en = false,
+	    .pull_down_en = false,
+	    .intr_type = GPIO_INTR_NEGEDGE,
+	};
+	ESP_ERROR_CHECK(gpio_config(&pincfg));
+
+	ESP_INTR_DISABLE(31);
+	intr_matrix_set(1, ETS_GPIO_INTR_SOURCE, 31);  // Assembler ISR aktivieren
+	ESP_INTR_ENABLE(31);
+
+#ifdef PIN_NUM_DEBUG_COPY
+	// noch ein Pin zur Fehlersuche: das zeigt an, wenn das Programm Daten von dem SPI Puffer in den VGA Puffer kopiert.
+	gpio_config_t pincfg2 =
+	{
+	    .pin_bit_mask = 1ULL << PIN_NUM_DEBUG_COPY,
+	    .mode = GPIO_MODE_OUTPUT,
+	    .pull_up_en = false,
+	    .pull_down_en = false,
+	};
+	ESP_ERROR_CHECK(gpio_config(&pincfg2));
+#endif
+}
+
+
+void IRAM_ATTR capture_task(void*)
+{
+	setup_abg();
+	volatile uint32_t* next = ABG_DMALIST;
+
+	while (1)
+	{
+#ifdef PIN_NUM_DEBUG_COPY
+		gpio_set_level(PIN_NUM_DEBUG_COPY,0);
+#endif
+
+		int a=1000000;
+		while ((next[0] & 0x80000000) != 0)  // darauf warten, dass der DMA-Kontroller den Transfer in den Puffer fertig meldet - der löscht das höchste Bit
+		{
+			if (bsyn_clock_last != bsyn_clock_diff && ABG_Scan_Line == 0)
+			{
+				// :200 Zeilen
+				// :240 MHz CPU-Takt
+				// *40 MHz Sample Rate
+				// = 1200
+				// :ABG_PIXEL_PER_LINE
+#ifdef PIN_NUM_DEBUG_COPY
+		gpio_set_level(PIN_NUM_DEBUG_COPY,1);
+#endif
+				bsyn_clock_last = bsyn_clock_diff;
+				double weite = bsyn_clock_diff / (1200.0f * ABG_PIXEL_PER_LINE);
+				double abst = 0.0f;
+				double step = modf((weite * ((double)(ABG_PIXEL_PER_LINE - BSYNC_PIXEL_ABSTAND))), &abst);
+				BSYNC_SAMPLE_ABSTAND = (int)abst;
+				int istep = round(step); 
+				int last = istep;
+				for (int i=0;i<640;i++)
+				{
+					step += weite;
+					istep = round(step);
+					PIXEL_STEP_LIST[i] = istep - last;
+					last = istep;
+				}
+
+				// :200 Zeilen
+				// :240 MHz CPU-Takt
+				// *40 MHz Sample Rate
+				// *8 Bit per Sample
+				// = 150
+				int spi_len = (bsyn_clock_diff / 150) - 600;
+				REG_SET_FIELD(SPI_MS_DLEN_REG(2), SPI_MS_DATA_BITLEN, spi_len);
+				REG_SET_BIT(SPI_CMD_REG(2), SPI_UPDATE);
+				ABG_RUN = true;
+#ifdef PIN_NUM_DEBUG_COPY
+		gpio_set_level(PIN_NUM_DEBUG_COPY,0);
+#endif
+			}
+
+			if (a==0)						 // Die Wartezeit ist abgelaufen!
+			{
+				for (int b=0;b<640*400;b++)  // VGA-Puffer leeren
+				{
+				    VGA_BUF[b]=0;
+				}
+				a=1000000;
+				ABG_RUN = false;
+			}
+			a--;
+		}
+
+#ifdef PIN_NUM_DEBUG_COPY
+		gpio_set_level(PIN_NUM_DEBUG_COPY,1);
+#endif
+		// Anzahl der empfangenen Bytes, wird vom DMA-Kontroller gesetzt
+		int b=(next[0] & 0xfff000) >> 12;
+
+		int sync = 0;
+		uint8_t* buf = (uint8_t*)next[1];
+
+		if ((buf[BSYNC_SUCHE_START] & 0x04)!=0x04)
+		{
+			BSYNC_SUCHE_START = BSYNC_SAMPLE_ABSTAND+1;
+		}
+
+		// Wir suchen den nächsten BSYNC-Impuls
+		for (int a=BSYNC_SUCHE_START; a<b; a++)
+		{
+			if ((buf[a] & 0x04)!=0x04)
+			{
+				sync = a;  // gefunden!
+				BSYNC_SUCHE_START = a - 20;
+				break;
+			}
+		}
+
+		uint32_t line = next[3];
+
+		// und nun die Pixel in den VGA-Puffer kopieren
+		if (sync>0 && line>=ABG_START_LINE && line<=ABG_START_LINE+399)
+		{
+            uint8_t* bufpos = (uint8_t*)((sync - BSYNC_SAMPLE_ABSTAND) + (int)buf);
+            uint8_t* vgapos = (uint8_t*)((line - ABG_START_LINE)*640 + (int)VGA_BUF);
+            for (int i=0;i<640;i++)
+            {
+                *vgapos = _STATIC_SYS_VALS[ACTIVESYS].colors[(*bufpos) & 3];
+                vgapos++;
+                bufpos+=PIXEL_STEP_LIST[i];
+            }
+		}
+		else
+		{
+			BSYNC_SUCHE_START = BSYNC_SAMPLE_ABSTAND+1;
+		}
+
+		// Puffer wieder für den DMA-Kontroller freigeben
+		next[0]=0x80ffffff;
+
+		// und zum nächsten Puffer springen
+		next = (uint32_t*)next[2];
+	}
+}
